1. Use Observable from RxJava (replacing callback) for the sake of syntactic sugar 2. Include dependency: rxjava and rxandroid 3. Remove Story model, now using ItemPOJO
