+++ app/src/main/java/io/pumpkinz/pumpkinreader/NewsListFragment.java
@@ -4,20 +4,32 @@ import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import io.pumpkinz.pumpkinreader.data.NewsAdapter;
 import io.pumpkinz.pumpkinreader.etc.DividerItemDecoration;
+import io.pumpkinz.pumpkinreader.model.ItemPOJO;
+import io.pumpkinz.pumpkinreader.rest.RestClient;
+import retrofit.Callback;
+import retrofit.RetrofitError;
+import retrofit.client.Response;
+import rx.Observable;
+import rx.functions.Action1;
+import rx.functions.Func1;
 
 
 public class NewsListFragment extends Fragment {
 
     RecyclerView newsList;
+    List<String> dataset;
+    NewsAdapter newsAdapter;
 
     public NewsListFragment() {
     }
@@ -38,7 +50,7 @@ public class NewsListFragment extends Fragment {
         layoutManager.setOrientation(LinearLayoutManager.VERTICAL);
         this.newsList.setLayoutManager(layoutManager);
 
-        List<String> dataset = Arrays.asList(
+        dataset = Arrays.asList(
                 "The Three Great Virtues of a Programmer: Laziness, Impatience, and Hubris",
                 "Philanthropy for Hackers",
                 "Ello mocks Facebook by being creepy",
@@ -49,11 +61,41 @@ public class NewsListFragment extends Fragment {
                 "Ask HN: How big does an open-source project need to be for a lifestyle business?",
                 "Fighting spam with Haskell"
         );
-        this.newsList.setAdapter(new NewsAdapter(dataset));
+        newsAdapter = new NewsAdapter(dataset);
+        this.newsList.setAdapter(newsAdapter);
 
         RecyclerView.ItemDecoration itemDecoration =
                 new DividerItemDecoration(getActivity(), DividerItemDecoration.VERTICAL_LIST);
         this.newsList.addItemDecoration(itemDecoration);
+
+        getStories();
     }
 
+    private void getStories() {
+        RestClient.service().listTopStories()
+                .flatMap(new Func1<List<Integer>, Observable<Integer>>() {
+                    @Override
+                    public Observable<Integer> call(List<Integer> integers) {
+                        Log.d("integer", integers.toString());
+                        return Observable.from(integers);
+                    }
+                })
+                .take(10)
+                .flatMap(new Func1<Integer, Observable<ItemPOJO>>() {
+                    @Override
+                    public Observable<ItemPOJO> call(Integer integer) {
+                        Log.d("integer", integer.toString());
+                        return RestClient.service().getItem(integer);
+                    }
+                })
+                .toList()
+                .subscribe(new Action1<List<ItemPOJO>>() {
+                    @Override
+                    public void call(List<ItemPOJO> itemPOJOs) {
+                        for (ItemPOJO item : itemPOJOs) {
+                            Log.d("topstories", item.toString());
+                        }
+                    }
+                });
+    }
 }
